ibmec-ecommerce-bot-main/requirements.txt
aiohttp>=3.10.0,<4.0.0
botbuilder-integration-aiohttp>=4.14.0
botbuilder-core>=4.14.0
botbuilder-dialogs>=4.14.0
requests>=2.23.0,<2.26.0
python-dotenv>=1.0.0


